+++ foxes/input/yaml/windio/read_outputs.py
@@ -1,3 +1,5 @@
+import numpy as np
+
 from foxes.utils import Dict
 import foxes.variables as FV
 import foxes.constants as FC
@@ -7,13 +9,13 @@ from .read_fields import foxes2wio
 
 def _read_turbine_outputs(wio_outs, olist, verbosity):
     """Reads the turbine outputs request"""
-    if "turbine_outputs" in wio_outs and wio_outs["turbine_outputs"].get(
+    if "turbine_outputs" in wio_outs and wio_outs["turbine_outputs"].get_item(
         "report", True
     ):
         turbine_outputs = Dict(
             wio_outs["turbine_outputs"], name=wio_outs.name + ".turbine_outputs"
         )
-        turbine_nc_filename = turbine_outputs.pop(
+        turbine_nc_filename = turbine_outputs.pop_item(
             "turbine_nc_filename", "turbine_outputs.nc"
         )
         output_variables = turbine_outputs["output_variables"]
@@ -52,19 +54,23 @@ def _read_turbine_outputs(wio_outs, olist, verbosity):
 
 def _read_flow_field(wio_outs, olist, verbosity):
     """Reads the flow field request"""
-    if "flow_field" in wio_outs and wio_outs["flow_field"].get("report", True):
+    if "flow_field" in wio_outs and wio_outs["flow_field"].get_item("report", True):
         flow_field = Dict(wio_outs["flow_field"], name=wio_outs.name + ".flow_field")
-        flow_nc_filename = flow_field.pop("flow_nc_filename", "flow_field.nc")
-        output_variables = flow_field.pop("output_variables")
-        z_planes = Dict(flow_field.pop("z_planes"), name=flow_field.name + ".z_planes")
+        flow_nc_filename = flow_field.pop_item("flow_nc_filename", "flow_field.nc")
+        output_variables = flow_field.pop_item("output_variables")
+
+        cases_run = Dict(
+            flow_field.pop_item("cases_run", {}), name=flow_field.name + ".cases_run"
+        )
+        if cases_run.pop_item("all_occurences"):
+            states_isel = None
+        else:
+            states_isel = cases_run.pop_item("occurences_list")
+
+        z_planes = Dict(flow_field.pop_item("z_planes"), name=flow_field.name + ".z_planes")
         z_sampling = z_planes["z_sampling"]
         xy_sampling = z_planes["xy_sampling"]
-        cases_run = Dict(
-            flow_field.pop("cases_run", {}), name=flow_field.name + ".cases_run"
-        )
-        states_isel = cases_run.get("subset", None)
-        if "all_occurences" in cases_run and cases_run.pop("all_occurences"):
-            states_isel = None
+
         if verbosity > 2:
             print("      Reading flow_field")
             print("        File name       :", flow_nc_filename)
@@ -78,16 +84,21 @@ def _read_flow_field(wio_outs, olist, verbosity):
             wind_direction=FV.WD,
         )
 
-        if z_sampling in ["hub_height", "default"]:
-            z = None
-        elif isinstance(z_sampling, (int, float)):
-            z = z_sampling
+        z_list = []
+        if z_sampling == "plane_list":
+            z_list = z_planes.pop_item("z_list")
+        elif z_sampling == "grid":
+            zb = z_planes.pop_item("z_bounds")
+            assert len(zb) == 2, f"Expecting two entries for z_bounds, got {zb}"
+            zn = z_planes.pop_item("z_number")
+            z_list = np.linspace(zb[0], zb[1], zn)
         else:
             raise NotImplementedError(
-                f"z_sampling '{z_sampling}' of type '{type(z_sampling).__name__}' is not supported (yet). Please give 'hub_height', 'default' or a float."
+                f"z_sampling '{z_sampling}' is not supported. Choices: 'plane_list', 'grid'."
             )
+        print("          z_list        :", list(z_list))
 
-        if xy_sampling == "default":
+        if xy_sampling == "grid":
             olist.append(Dict(
                 output_type="SliceData",
                 verbosity_delta=3,
@@ -107,7 +118,7 @@ def _read_flow_field(wio_outs, olist, verbosity):
             ))
         else:
             raise NotImplementedError(
-                f"xy_sampling '{xy_sampling}' is not supported (yet)"
+                f"xy_sampling '{xy_sampling}' is not supported. Choices: 'grid'."
             )
 
 
@@ -132,7 +143,7 @@ def read_outputs(wio_outs, olist, verbosity=1):
     :group: input.yaml.windio
 
     """
-    odir = wio_outs.pop("output_folder", ".")
+    odir = wio_outs.pop_item("output_folder", ".")
     if verbosity > 2:
         print("  Reading outputs")
         print("    Output dir:", odir)