+++ foxes/output/__init__.py
@@ -13,6 +13,7 @@ from .turbine_type_curves import TurbineTypeCurves
 from .animation import Animator
 from .calc_points import PointCalculator
 from .slice_data import SliceData
+from .slices_data import SlicesData
 from .rotor_point_plots import RotorPointPlot
 from .state_turbine_table import StateTurbineTable
 from .plt import plt