+++ CHANGELOG.md
@@ -702,6 +702,8 @@ This major version introduces the concept of `Engines` which handle the chunking
   - Now three directories are defined: `work_dir`, `input_dir`, `output_dir`. If not defined, falling back to `work_dir`, defaulting to cwd.
 - Inputs:
   - Removing `SliceDataNC`, since now `FieldDataNC` can handle data without height dependency
+- Outputs:
+  - New output `SlicesData` which combines slices from `SliceData` into a single NetCDF file, resulting in full 3d field output (or 4d, with time/states dimension)
 - Models:
   - New point model `Ustar2TI`, estimates TI from friction velocity